luci-app-overthebox: Move tc_stats() from lib
