+++ overthebox/files/usr/lib/lua/overthebox.lua
@@ -719,73 +719,6 @@ function lock(name)
   return file
 end
 
-function tc_stats()
-  local output = {}
-  for line in string.gmatch((sys.exec("tc -s q")), '[^\r\n]+') do
-    table.insert(output, line)
-  end
-
-
-  local result = {}
-  result["upload"] = {}
-  local curdev;
-  local curq;
-  for i=1, #output do
-    if string.byte(output[i]) ~= string.byte(' ') then
-      curdev = nil
-      curq = nil
-    end
-    if string.match(output[i], "dev ([^%s]+)") then
-      curdev = string.match(output[i], "dev ([^%s]+)")
-    end
-    if string.match(output[i], "sfq (%d+)") then
-      curq = string.match(output[i], "sfq (%d+)")
-    end
-    if curdev and curq then
-      for bytes, pkt, dropped, overlimits, reque in string.gmatch(output[i], "Sent (%d+) bytes (%d+) pkt %(dropped (%d+), overlimits (%d+) requeues (%d+)") do
-        -- print("["..curdev..", "..curq..", "..bytes.. ", "..pkt..", "..dropped..", "..overlimits..", ".. reque .. "]")
-        if result["upload"][curq] == nil then
-          result["upload"][curq] = {}
-        end
-        result["upload"][curq][curdev] = { bytes=bytes, pkt=pkt, dropped=dropped, overlimits=overlimits, requeues=reque }
-      end
-    end
-  end
-
-  output = {}
-  result["download"] = {}
-  local tcstats = json.decode(sys.exec("curl -s --max-time 1 api/qos/tcstats"))
-  if tcstats and tcstats.raw_output then
-
-    for line in string.gmatch(tcstats.raw_output, '[^\r\n]+') do
-      table.insert(output, line)
-    end
-
-    for i=1, #output do
-      if string.byte(output[i]) ~= string.byte(' ') then
-        curdev = nil
-        curq = nil
-      end
-      if string.match(output[i], "dev ([^%s]+)") then
-        curdev = string.match(output[i], "dev ([^%s]+)")
-      end
-      if string.match(output[i], "sfq (%d+)") then
-        curq = string.match(output[i], "sfq (%d+)")
-      end
-      if curdev and curq then
-        for bytes, pkt, dropped, overlimits, reque in string.gmatch(output[i], "Sent (%d+) bytes (%d+) pkt %(dropped (%d+), overlimits (%d+) requeues (%d+)") do
-          -- print("["..curdev..", "..curq..", "..bytes.. ", "..pkt..", "..dropped..", "..overlimits..", ".. reque .. "]")
-          if result["download"][curq] == nil then
-            result["download"][curq] = {}
-          end
-          result["download"][curq][curdev] = { bytes=bytes, pkt=pkt, dropped=dropped, overlimits=overlimits, requeues=reque }
-        end
-      end
-    end
-  end
-  return result
-end
-
 function log(msg)
   posix.syslog( posix.LOG_INFO, msg)
 end
